+++ hubspot_service/internal/adapters/outbound/asynq/enqueuer.go
@@ -104,11 +104,11 @@ func (e *Enqueuer) EnqueueUpsertAsesor(ctx context.Context, a domain.AsesorPaylo
 // ---------- Payloads (ser/deser-able JSON) ----------
 
 type SyncResultPayload struct {
-	DNI           string `json:"dni"`
-	ExamTypeCode  string `json:"exam_type_code"`
-	Score         int32  `json:"score"`
-	MaxScore      int32  `json:"max_score"`
-	AttemptID     string `json:"attempt_id"`
+	DNI           string  `json:"dni"`
+	ExamTypeCode  string  `json:"exam_type_code"`
+	Score         float64 `json:"score"`
+	MaxScore      float64 `json:"max_score"`
+	AttemptID     string  `json:"attempt_id"`
 	ContactRecord string `json:"contact_record_id,omitempty"`
 	SubmittedAt   string `json:"submitted_at"`
 }
